simplify `Sheet` tire block and init `SheetData` sub-block
remove `/api/sheets/[token]` route and `useSheetData` hook
